AjaxSearchController
