--- domain-module/src/main/java/io/manco/maxim/sbmm/domain/OperationParameters.java
@@ -1,18 +1,0 @@
-package io.manco.maxim.sbmm.domain;
-
-public class OperationParameters {
-  private String name;
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  @Override
-  public String toString() {
-    return name;
-  }
-}
