+++ domain-module/src/main/java/io/manco/maxim/sbmm/domain/WatchListDesc.java
@@ -160,4 +160,21 @@ public class WatchListDesc {
             + "\n " + "market data freq = " + this.getMarketDataFrequency() + "\n " + "data set name = "
             + this.getWatchListName() + "\n " + "data set description = " + this.getWatchListDetails() + "\n ");
   }
+  
+  public static class OperationParameters{
+    private String name;
+
+    public String getName() {
+      return name;
+    }
+
+    public void setName(String name) {
+      this.name = name;
+    }
+    
+    @Override
+    public String toString() {
+      return name;
+    }
+  }
 }
